+++ src/javadoc/diskmgr/DB.html
@@ -60,7 +60,7 @@ Class  DB</H2>
 <PRE>
 java.lang.Object
   |
-  +--<B>diskmgr.DB</B>
+  +--<B>diskmgr.bigDB</B>
 </PRE>
 <DL>
 <DT><B>All Implemented Interfaces:</B> <DD><A HREF="../global/GlobalConst.html">GlobalConst</A></DD>
